move discretization logic to utils
